+++ src/binbuf.h
@@ -28,6 +28,7 @@ struct binbuf {
 };
 
 void binbuf_init(struct binbuf *sb);
+char *binbuf_cpy(struct binbuf *dest, const struct binbuf *src);
 void binbuf_free(struct binbuf *sb);
 
 #endif /* ifndef _BINBUF_H */